requirements.txt
wheel
Pillow
python-dotenv
tweepy
black
requests